feat: refactor chat
